Stack and multiprocessor alive!

+++ arp/ip/ac_tlm_lock/ac_tlm_lock.h
@@ -116,7 +116,7 @@ public:
 
 
 private:
-  uint8_t memory;
+  uint32_t memory;
 
 };
